Refactored p_2,  Added input validation
