+++ tests/test_p2/__init__.py
@@ -1,0 +1,1 @@
+"""p2 test module."""
