+++ tests/test_p2/test_p_2_dataframe.py
@@ -1,6 +1,6 @@
 import pandas as pd
 
-from pytry.p_2_dataframe import score_board
+from pytry.p_2.p_2_dataframe import score_board
 
 
 def test_get_other_side():
@@ -15,20 +15,20 @@ def test_update_scores():
     )
 
     game = {"A": 1, "B": 1}
-    scores = score_board.update_scores(scores, game, "A", "B")
+    scores = score_board.update_country_scores(scores, game, "A", "B")
 
     assert scores.at["A", "draws"] == 1
     assert scores.at["A", "points"] == 1
     assert scores.at["A", "goal_difference"] == 0
 
     game = {"A": 2, "C": 1}
-    scores = score_board.update_scores(scores, game, "A", "C")
+    scores = score_board.update_country_scores(scores, game, "A", "C")
     assert scores.at["A", "wins"] == 1
     assert scores.at["A", "points"] == 4
     assert scores.at["A", "goal_difference"] == 1
 
     game = {"A": 1, "D": 3}
-    scores = score_board.update_scores(scores, game, "A", "D")
+    scores = score_board.update_country_scores(scores, game, "A", "D")
     assert scores.at["A", "loses"] == 1
     assert scores.at["A", "points"] == 4
     assert scores.at["A", "goal_difference"] == -1
